rename pbp to pbp-hmb-gen
