+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 2024-08
 
+- 1.4.2:
+    - renamed `pbp` CLI program to `pbp-hmb-gen` 
+
 - 1.4.1 - HMB generation CLI now with `--s3-unsigned` option to use unsigned config.
 
 - 1.4.0 with improved "meta-gen" functionality
