+++ justfile
@@ -32,7 +32,7 @@ tgz:
 
 # Run main (on gizo)
 main-gizo date="20220902" output_dir="/PAM_Analysis/pypam-space/test_output/daily":
-    PYTHONPATH=. poetry run python pbp/main.py \
+    PYTHONPATH=. poetry run python pbp/main_hmb_generator.py \
                  --json-base-dir=json \
                  --date={{date}} \
                  --voltage-multiplier=3 \
@@ -43,7 +43,7 @@ main-gizo date="20220902" output_dir="/PAM_Analysis/pypam-space/test_output/dail
 
 # Run main (on gizo) with some initial test jsons
 main-gizo-test *more_args="":
-    PYTHONPATH=. poetry run python pbp/main.py \
+    PYTHONPATH=. poetry run python pbp/main_hmb_generator.py \
                  --json-base-dir=tests/json \
                  --date=20220902 \
                  --voltage-multiplier=3 \
@@ -67,7 +67,7 @@ main-gizo-multiple-days year month *days="":
       base="$output_dir/milli_psd_$date"
       out="$base.out"
       echo "running: day=$day output_dir=$output_dir"
-      poetry run python pbp/main.py \
+      poetry run python pbp/main_hmb_generator.py \
              --json-base-dir=json \
              --date="$date" \
              --voltage-multiplier=3 \
@@ -87,7 +87,7 @@ main-mb05 *more_args="":
     mkdir -p NB_SPACE/OUTPUT
 
     PYTHONPATH=. EXCLUDE_LOG_TIME=yes \
-        poetry run python pbp/main.py \
+        poetry run python pbp/main_hmb_generator.py \
                  --date=20220812 \
                  --json-base-dir=NB_SPACE/JSON \
                  --voltage-multiplier=1 \
@@ -113,7 +113,7 @@ main-nrs11 date='20200101' *more_args='':
     mkdir -p $WS/DOWNLOADS
     mkdir -p $WS/OUTPUT
     PYTHONPATH=. EXCLUDE_LOG_TIME=yes \
-        poetry run python pbp/main.py \
+        poetry run python pbp/main_hmb_generator.py \
                  --date={{date}} \
                  --gs \
                  --json-base-dir=$WS/noaa-passive-bioacoustic_nrs_11_2019-2021 \
@@ -139,7 +139,7 @@ main-nrs11-multiple-days year month *days="":
     export PYTHONPATH=.
     for day in {{days}}; do
       date=$(printf "%04d%02d%02d" {{year}} {{month}} "$day")
-      poetry run python pbp/main.py \
+      poetry run python pbp/main_hmb_generator.py \
              --date="$date" \
              --gs \
              --json-base-dir=$WS/noaa-passive-bioacoustic_nrs_11_2019-2021 \
@@ -231,7 +231,7 @@ main-cloud-chumash-basic-test date="20230108":
 
 # Run main
 main *args="":
-    PYTHONPATH=. poetry run python pbp/main.py {{args}}
+    PYTHONPATH=. poetry run python pbp/main_hmb_generator.py {{args}}
 
 ##############
 # misc/utils:
